Sonora concluído

+++ public/js/acesso_sonora.js
@@ -5,26 +5,13 @@ fetch("/sonora/listar", {
         resposta
             .json()
             .then(json => {
+                localStorage.ARTISTAS = JSON.stringify(json);
                 exibirListaArtistas(json);
             });
     }
 }).catch(err => {
     console.log(err);
 });
-
-/*
-fetch("/sonora/listar/pais", {
-    method: "GET"
-}).then(resposta => {
-    if (resposta.ok) {
-        resposta
-            .json()
-            .then(json => {
-                console.log(json);
-            });
-    }
-});
-*/
 
 function exibirListaArtistas(listaArtistas = [{}]) {
     const corpoTabela = document.querySelector("tbody");
@@ -36,33 +23,53 @@ function exibirListaArtistas(listaArtistas = [{}]) {
     listaArtistas.forEach(artista => {
         corpoTabela.innerHTML += `
             <tr>
-                <!-- <td></td> -->
                 <td>${artista.id}</td>
-                <td>${artista.nome}</td>
-                <td>${decidirVazio(artista.qtdLineups)}</td>
-                <td>${decidirVazio(artista.qtdDadosLastFm)}</td>
-                <td>${decidirVazio(artista.qtdDadosSpotify)}</td>
+                <td class="celula-texto">${artista.nome}</td>
+                <td>${artista.qtdLineups/2}</td>
+                <td>${artista.qtdDadosLastFm}</td>
+                <td>${artista.qtdDadosSpotify}</td>
+                <td class="celula-acao" onclick="mostrarModal('Editar', ${artista.id})">Editar</td>
+                <td class="celula-acao" onclick="executarDelete(${artista.id})">Excluir</td>
             </tr>
         `
     });
 }
 
-function decidirVazio(numero = 0) {
-    if (numero == 0) {
-        return numero = ''
+function mostrarModal(operacao = '', idArtista = 0) {
+    const inpModal = document.querySelector("#inp_nome_artista");
+    
+    if (operacao == 'Editar') {
+        const artista = JSON.parse(localStorage.ARTISTAS).find(artista => artista.id === idArtista);
+
+        inpModal.value = artista.nome
+        inpModal.setAttribute("name", artista.id)
+    } else {
+        inpModal.value = "";
     }
 
-    return numero;
+    document.querySelector("#modal span").innerHTML = operacao;
+    document.querySelector("#modal").style.display = "block";
 }
 
-document.querySelector("#legenda_tabela>button").addEventListener("click", e => {
-    const infos = {
-        nome: "Novo artista",
-        idRelacionado: 1
-    };
+function cancelarModal() {
+    document.querySelector("#modal").style.display = "none";
+}
 
-    registrarNovoArtista(infos);
-});
+function enviarRequisicao() {
+    const inpNome = document.querySelector("#inp_nome_artista");
+
+    const infos = {
+        id: inpNome.name,
+        nome: inpNome.value,
+        idRelacionado: null
+    }
+
+    if (inpNome.hasAttribute("name")) {
+        registrarEditarArtista(infos);
+    } else {
+        registrarNovoArtista(infos);
+    }
+}
 
 function registrarNovoArtista(infos = {}) {   
     fetch("/sonora/criar", {
@@ -72,24 +79,16 @@ function registrarNovoArtista(infos = {}) {
         },
         body: JSON.stringify(infos) 
     }).then(resposta => {
-        if (resposta.ok) alert("DEU BOM");
-        else alert("slk checa ai");
+        if (resposta.ok) {
+            alert("Artista registrado!");
+            location.reload();
+        } else {
+            alert("Houve um erro!");
+        }   
     }).catch(err => {
-        console.log(err)
+        alert("Houve um erro!")
     });
 }
-
-/*
-document.querySelector("#btn_teste").addEventListener("click", e => {
-    const infos = {
-        id: '3',
-        nome: 'nome do caboco',
-        idRelacionado: '1'
-    };
-    
-    registrarEditarArtista(infos);
-});
-*/
 
 function registrarEditarArtista(infos = {}) {
     fetch("/sonora/editar", {
@@ -99,24 +98,33 @@ function registrarEditarArtista(infos = {}) {
         },
         body: JSON.stringify(infos)
     }).then(resposta => {
-        if (resposta.ok) alert("deu bom");
-        else alert("slk checa ai");
+        if (resposta.ok) {
+            alert("Artista editado");
+            location.reload();
+        } else {
+            alert("Houve um erro!");
+        }
     }).catch(err => {
-        console.log(err);
+        alert("Houve um erro!");
     })
 }
 
-
-document.querySelector("#btn_teste").addEventListener("click", e => { 
-    executarDelete(4);
-});
-
 function executarDelete(idArtista = 0){
+    const exclusao = confirm("Tem certeza que deseja deletar o artista? Essa ação afetará todo o banco de dados e não poderá ser desfeita.")
+
+    if(!exclusao) {
+        return;
+    }
+
     fetch(`/sonora/deletar/${idArtista}`, {
         method: "DELETE"
     }).then(resposta => {
-        if (resposta.ok) alert("deu bom");
-        else alert("deu ruim");
+        if (resposta.ok) {
+            alert("Exclusão bem sucedida.")
+            location.reload();
+        } else {
+            alert("Houve um erro!")
+        } 
     }).catch(erro => {
         console.log(erro);
     });
